+++ emperor/support_files/emperor/js/emperor.js
@@ -14,6 +14,7 @@ var g_plotSpheres = {};
 var g_plotEllipses = {};
 var g_plotTaxa = {};
 var g_plotVectors = {};
+var g_parallelPlots = []
 
 // sample identifiers of all items that are plotted
 var g_plotIds = [];
@@ -81,7 +82,7 @@ function resetCamera() {
 	// We need to reset the camera controls first before modifying the values of the camera (this is the reset view!)
 	g_sceneControl.reset();
 	
-	g_sceneCamera.aspect = document.getElementById('main_plot').offsetWidth/document.getElementById('main_plot').offsetHeight;
+	g_sceneCamera.aspect = document.getElementById('pcoaPlotWrapper').offsetWidth/document.getElementById('pcoaPlotWrapper').offsetHeight;
 	g_sceneCamera.rotation.set( 0, 0, 0);
 	g_sceneCamera.updateProjectionMatrix();
 	
@@ -254,30 +255,30 @@ function toggleContinuousAndDiscreteColors(element){
   coloring scheme that the system is currently using (discrete or continuous).
 */
 function getColorList(vals) {
-	var colors = [];
+	var colors = {};
 
 	// cases with one or two categories are basically the same no matter if the
 	// coloring scheme is continuous or discrete; choose red or red and blue
 	if(vals.length == 1){
-		colors[0] = new THREE.Color();
-		colors[0].setHex("0xff0000");
+		colors[vals[0]] = new THREE.Color();
+		colors[vals[0]].setHex("0xff0000");
 	}
 	else if (vals.length == 2) {
-		colors[0] = new THREE.Color();
-		colors[0].setHex("0xff0000");
-		colors[1] = new THREE.Color();
-		colors[1].setHex("0x0000ff");
+		colors[vals[0]] = new THREE.Color();
+		colors[vals[0]].setHex("0xff0000");
+		colors[vals[1]] = new THREE.Color();
+		colors[vals[1]].setHex("0x0000ff");
 	}
 	else {
 		for(var index in vals){
-			colors[index] = new THREE.Color();
+			colors[vals[index]] = new THREE.Color();
 			if(g_useDiscreteColors){
 				// get the next available color
-				colors[index].setHex(getDiscreteColor(index)*1);
+				colors[vals[index]].setHex(getDiscreteColor(index)*1);
 			}
 			else{
 				// multiplying the value by 0.66 makes the colormap go R->G->B
-				THREE.ColorConverter.setHSV(colors[index], index*.66/vals.length, 1, 1)
+				THREE.ColorConverter.setHSV(colors[vals[index]], index*.66/vals.length, 1, 1)
 			}
 		}
 	}
@@ -352,7 +353,7 @@ function colorByMenuChanged() {
 
 	vals = _splitAndSortNumericAndAlpha(dedupe(vals));
 	colors = getColorList(vals);
-
+	
 	// build the colorby table in HTML
 	var lines = "<table>";
 	for(var i in vals){
@@ -382,10 +383,10 @@ function colorByMenuChanged() {
 		var idString = "r"+i+"c"+g_categoryIndex;
 
 		// get the div built earlier and turn it into a color picker
-		$('#'+idString).css('backgroundColor',"#"+colors[i].getHexString());
+		$('#'+idString).css('backgroundColor',"#"+colors[vals[i]].getHexString());
 		$("#"+idString).spectrum({
 			localStorageKey: 'key',
-			color: colors[i].getHexString(),
+			color: colors[vals[i]].getHexString(),
 			showInitial: true,
 			showInput: true,
 			change:
@@ -396,11 +397,47 @@ function colorByMenuChanged() {
 						c = "#"+c.charAt(1)+c.charAt(1)+c.charAt(2)+c.charAt(2)+c.charAt(3)+c.charAt(3);
 					}
 					colorChanged($(this).attr('name'), c);
+					colors[$(this).attr('name')] = c;
+					colorParallelPlots(vals, colors);
 				}
 		});
 	}
 
+	colorParallelPlots(vals, colors);
 	setKey(vals, colors);
+}
+
+function colorParallelPlots(vals,colors) 
+{
+	pwidth = document.getElementById('parallelPlotWrapper').offsetWidth
+	pheight = document.getElementById('parallelPlotWrapper').offsetHeight
+	
+	document.getElementById('parallelPlotWrapper').innerHTML = '<div id="parallelPlot" class="parcoords" style="width:'+pwidth+'px;height:'+pheight+'px"></div>'
+	
+	var color = function(d) {
+		var sid = d[0];
+		var divid = sid.replace(/\./g,'')+"_key";
+		var catValue = g_mappingFileData[sid][g_categoryIndex];
+		var catColor = colors[catValue];
+
+		try {
+			var hex = '#'+catColor.getHexString();
+		}catch(TypeError) {
+			var hex = catColor;
+		}
+		return hex;
+	}
+	
+	var pc = d3.parcoords()("#parallelPlot")
+	  .data(g_parallelPlots)
+	  .color(color)
+	  .margin({ top: 40, left: 50, bottom: 40, right: 0 })
+	  .mode("queue")
+	  .render()
+	  .brushable();
+	  
+	$('.parcoords text').css('stroke', $('#parallelaxeslabelcolor').css('backgroundColor'));
+	$('.parcoords .axis line, .parcoords .axis path').css('stroke', $('#parallelaxescolor').css('backgroundColor'));
 }
 
 /*Callback when the scaling by drop-down menu changes
@@ -601,7 +638,7 @@ function toggleVisible(value) {
 function setKey(values, colors) {
 	if(g_keyBuilt){
 		for(var i = 0; i < values.length; i++){
-			colorChanged(values[i], '#'+colors[i].getHexString());
+			colorChanged(values[i], '#'+colors[values[i]].getHexString());
 		}
 	}
 	else {
@@ -610,7 +647,7 @@ function setKey(values, colors) {
 			var sid = g_plotIds[i];
 			var divid = sid.replace(/\./g,'')+"_key";
 			var catValue = g_mappingFileData[sid][g_categoryIndex];
-			var catColor = colors[values.indexOf(catValue)];
+			var catColor = colors[catValue];
 			keyHTML += "<tr id=\""+divid+"row\"><td><div id=\""+divid+"\" name=\""+sid+"\" class=\"colorbox\" style=\"background-color:#";
 			keyHTML += catColor.getHexString();
 			keyHTML += ";\"></div>";
@@ -754,11 +791,11 @@ function labelMenuChanged() {
 		var idString = "r"+i+"c"+g_categoryIndex;
 
 		// get the div built earlier and turn it into a color picker
-		$('#'+idString+'Label').css('backgroundColor',"#"+colors[i].getHexString());
-		labelColorChanged(vals[i], "#"+colors[i].getHexString());
+		$('#'+idString+'Label').css('backgroundColor',"#"+colors[vals[i]].getHexString());
+		labelColorChanged(vals[i], "#"+colors[vals[i]].getHexString());
 
 		$("#"+idString+'Label').spectrum({
-			color: colors[i].getHexString(),
+			color: colors[vals[i]].getHexString(),
 			showInitial: true,
 			showPalette: true,
 			palette: [['red', 'green', 'blue']],
@@ -1016,6 +1053,9 @@ function sphereRadiusChange(ui, category) {
 /*Setup the interface elements required for the sidebar of the main interface*/
 function setJqueryUi() {
 	$("#menutabs").tabs();
+	$("#plottype").buttonset();
+	$("input[name='plottype']").change(togglePlots);
+	
 	$("#labelColor").css('backgroundColor', '#fff');
 
 	$("#labelColor").spectrum({
@@ -1133,8 +1173,62 @@ function setJqueryUi() {
 	});
 	document.getElementById('labelopacity').innerHTML = $( "#lopacityslider" ).slider( "value")+"%"
 
+	
+	//default color for parallel plots axes label is white
+	$('#parallelaxeslabelcolor').css('backgroundColor',"#ffffff");
+	$("#parallelaxeslabelcolor").spectrum({
+		localStorageKey: 'key',
+		color: "#ffffff",
+		showInitial: true,
+		showInput: true,
+		showPalette: true,
+		preferredFormat: "hex6",
+		palette: [['white', 'black']],
+		change:
+			function(color) {
+				// pass a boolean flag to convert to hex6 string
+				var c = color.toHexString(true);
+
+				// create a new three.color from the string
+				var parallelAxesLabelColor = new THREE.Color();
+				parallelAxesLabelColor.setHex(c.replace('#','0x'));
+
+				// set the color for the box and for the renderer
+				$(this).css('backgroundColor', c);
+				//set css for the lines...
+				$('.parcoords text').css('stroke', c);
+			}
+	});
+
+	//default color for parallel plots axes is white
+	$('#parallelaxescolor').css('backgroundColor',"#ffffff");
+	$("#parallelaxescolor").spectrum({
+		localStorageKey: 'key',
+		color: "#ffffff",
+		showInitial: true,
+		showInput: true,
+		showPalette: true,
+		preferredFormat: "hex6",
+		palette: [['white', 'black']],
+		change:
+			function(color) {
+				// pass a boolean flag to convert to hex6 string
+				var c = color.toHexString(true);
+
+				// create a new three.color from the string
+				var parallelAxesColor = new THREE.Color();
+				parallelAxesColor.setHex(c.replace('#','0x'));
+
+				// set the color for the box and for the renderer
+				$(this).css('backgroundColor', c);
+				//set css for the lines...
+				$('.parcoords .axis line, .parcoords .axis path').css('stroke', c);
+			}
+	});
+
 	// the default color palette for the background is black and white
 	$('#rendererbackgroundcolor').css('backgroundColor',"#000000");
+	$('#parallelPlotWrapper').css('backgroundColor',"#000000");
 	$("#rendererbackgroundcolor").spectrum({
 		localStorageKey: 'key',
 		color: "#000000",
@@ -1155,6 +1249,8 @@ function setJqueryUi() {
 				// set the color for the box and for the renderer
 				$(this).css('backgroundColor', c);
 				g_mainRenderer.setClearColor(rendererBackgroundColor, 1);
+				
+				$('#parallelPlotWrapper').css('backgroundColor', c);
 			}
 	});
 }
@@ -1432,6 +1528,54 @@ function clean_label_refresh_axes() {
 	$("#refresh_axes_label").html("");
 }
 
+function togglePlots() {
+	if(document.getElementById('pcoa').checked)
+	{
+		document.getElementById('pcoaPlotWrapper').className = document.getElementById('pcoaPlotWrapper').className.replace(/(?:^|\s)invisible(?!\S)/ , '');
+		document.getElementById('pcoaoptions').className = document.getElementById('pcoaoptions').className.replace(/(?:^|\s)invisible(?!\S)/ , '');
+		document.getElementById('pcoaaxes').className = document.getElementById('pcoaaxes').className.replace(/(?:^|\s)invisible(?!\S)/ , '');
+	  	document.getElementById('parallelPlotWrapper').className += ' invisible'
+	  	document.getElementById('paralleloptions').className += ' invisible'
+	  	document.getElementById('parallelaxes').className += ' invisible'
+	  	$("#menutabs").tabs('select',0);
+	  	$("#menutabs").tabs({disabled: []});
+	}
+	else
+	{
+		document.getElementById('parallelPlotWrapper').className = document.getElementById('parallelPlotWrapper').className.replace(/(?:^|\s)invisible(?!\S)/ , '');
+		document.getElementById('paralleloptions').className = document.getElementById('paralleloptions').className.replace(/(?:^|\s)invisible(?!\S)/ , '');
+		document.getElementById('parallelaxes').className = document.getElementById('parallelaxes').className.replace(/(?:^|\s)invisible(?!\S)/ , '');
+	  	document.getElementById('pcoaPlotWrapper').className += ' invisible'
+	  	document.getElementById('pcoaoptions').className += ' invisible'
+	  	document.getElementById('pcoaaxes').className += ' invisible'
+	  	$("#menutabs").tabs('select',0);
+	  	$("#menutabs").tabs({disabled: [2,3,4]});
+	  	colorByMenuChanged();
+	}
+}
+
+function setParallelPlots() {	
+	g_parallelPlots = []
+	var num_axes = g_fractionExplained.length
+
+	for(p in g_spherePositions)
+	{
+		var dataline = []
+		
+		dataline.push(g_spherePositions[p].name)
+		for(var i = 1; i < num_axes+1; i++)
+		{
+			dataline.push(g_spherePositions[p]['P'+i])
+		}
+		g_parallelPlots.push(dataline)
+	}
+	
+	pwidth = document.getElementById('pcoaPlotWrapper').offsetWidth
+	pheight = document.getElementById('pcoaPlotWrapper').offsetHeight
+
+	document.getElementById('parallelPlotWrapper').innerHTML = '<div id="parallelPlot" class="parcoords" style="width:'+pwidth+'px;height:'+pheight+'px"></div>'
+}
+
 /*Setup and initialization function for the whole system
 
   This function will set all of the WebGL elements that are required to exist
@@ -1440,7 +1584,7 @@ function clean_label_refresh_axes() {
 */
 $(document).ready(function() {
 	setJqueryUi()
-
+	
 	// Detecting that webgl is activated
 	if ( ! Detector.webgl ) Detector.addGetWebGLMessage();
 
@@ -1448,12 +1592,12 @@ $(document).ready(function() {
 	var particles, geometry, parameters, i, h, color;
 	var mouseX = 0, mouseY = 0;
 
-	var winWidth = Math.min(document.getElementById('main_plot').offsetWidth,document.getElementById('main_plot').offsetHeight), view_angle = 35, view_near = 0.0000001, view_far = 10000;
-	var winAspect = document.getElementById('main_plot').offsetWidth/document.getElementById('main_plot').offsetHeight;
+	var winWidth = Math.min(document.getElementById('pcoaPlotWrapper').offsetWidth,document.getElementById('pcoaPlotWrapper').offsetHeight), view_angle = 35, view_near = 0.0000001, view_far = 10000;
+	var winAspect = document.getElementById('pcoaPlotWrapper').offsetWidth/document.getElementById('pcoaPlotWrapper').offsetHeight;
 
 	$(window).resize(function() {
-		winWidth = Math.min(document.getElementById('main_plot').offsetWidth,document.getElementById('main_plot').offsetHeight);
-		winAspect = document.getElementById('main_plot').offsetWidth/document.getElementById('main_plot').offsetHeight;
+		winWidth = Math.min(document.getElementById('pcoaPlotWrapper').offsetWidth,document.getElementById('pcoaPlotWrapper').offsetHeight);
+		winAspect = document.getElementById('pcoaPlotWrapper').offsetWidth/document.getElementById('pcoaPlotWrapper').offsetHeight;
 		g_sceneCamera.aspect = winAspect;
 		g_sceneCamera.updateProjectionMatrix();
 	});
@@ -1467,8 +1611,8 @@ $(document).ready(function() {
 		g_sceneCamera = new THREE.PerspectiveCamera(view_angle, winAspect, view_near, view_far);
 		g_sceneCamera.position.set(0, 0, 0);
 
-		$('#main_plot canvas').attr('width',document.getElementById('main_plot').offsetWidth);
-		$('#main_plot canvas').attr('height',document.getElementById('main_plot').offsetHeight);
+		$('#main_plot canvas').attr('width',document.getElementById('pcoaPlotWrapper').offsetWidth);
+		$('#main_plot canvas').attr('height',document.getElementById('pcoaPlotWrapper').offsetHeight);
 
 		g_mainScene = new THREE.Scene();
 		g_mainScene.fog = new THREE.FogExp2( 0x000000, 0.0009);
@@ -1514,10 +1658,14 @@ $(document).ready(function() {
 			$("#showbycombo").append(line);
 			$("#labelcombo").append(line);
 		}
-
+		
+		setParallelPlots();
+		
 		colorByMenuChanged();
 		showByMenuChanged();
 		scalingByMenuChanged();
+		
+		togglePlots();
 
 		// the light is attached to the camera to provide a 3d perspective
 		g_sceneLight = new THREE.DirectionalLight(0x999999, 2);
@@ -1542,7 +1690,8 @@ $(document).ready(function() {
 		// renderer, the default background color is black
 		g_mainRenderer = new THREE.WebGLRenderer({ antialias: true });
 		g_mainRenderer.setClearColor(rendererBackgroundColor, 1);
-		g_mainRenderer.setSize( document.getElementById('main_plot').offsetWidth, document.getElementById('main_plot').offsetHeight );
+		g_mainRenderer.setSize( document.getElementById('pcoaPlotWrapper').offsetWidth, document.getElementById('pcoaPlotWrapper').offsetHeight );
+		// g_mainRenderer.setSize( document.getElementById('vizualizations').offsetWidth , document.getElementById('vizualizations').offsetHeight );
 		g_mainRenderer.sortObjects = true;
 		main_plot.append(g_mainRenderer.domElement);
 
@@ -1700,7 +1849,7 @@ $(document).ready(function() {
    
 	function render() {
 		g_sceneControl.update();
-		g_mainRenderer.setSize( document.getElementById('main_plot').offsetWidth, document.getElementById('main_plot').offsetHeight );
+		g_mainRenderer.setSize( document.getElementById('pcoaPlotWrapper').offsetWidth, document.getElementById('pcoaPlotWrapper').offsetHeight );
 		g_mainRenderer.render( g_mainScene, g_sceneCamera);
 	}
 });