+++ ChangeLog.md
@@ -7,6 +7,7 @@ Emperor 0.9.1 (changes since Emperor 0.9.0 go here)
 * GUI is usable in mobile devices that support WebGL.
 * The user can select the number of axes to be considered in the GUI and re-plot using lower axes; this is, for example: PC3 vs PC4 vs PC10.
 * In missing_custom_axes_values you can reference other column within the mapping file to place the samples without numeric values at different points in the gradient.
+* Parallel plots functionality.
 
 *Bug Fixes*